Change tooling structure
